+++ client/src/widgets/TechInterviewRandomWidget/index.ts
@@ -1,0 +1,2 @@
+export { default as TechInterviewRandomWidget } from './TechInterviewRandomWidget'
+export { default } from './TechInterviewRandomWidget'